FAKE-NEWS-main/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
google-generativeai>=0.7.2
spacy>=3.7.0
httpx>=0.27.0
newspaper3k>=0.2.8
python-dotenv>=1.0.1
beautifulsoup4>=4.12.3
lxml_html_clean>=0.1.1
nltk>=3.9.1
